chore: move contracts and ui directories to packages directory
The contracts and ui directories have been moved to the packages directory. This change was necessary to improve the project structure and organization. The move includes a rename of all the files and directories within these directories. This will ensure that the contracts and ui components are properly encapsulated within the packages directory.