chore(contracts): update `prepare` script in `package.json`
The `prepare` script in the `package.json` file has been modified to change the directory in the `cd` command. The previous value was `cd ..` and it has been updated to `cd ../..`. This change was made to properly locate the `.husky` directory during the preparation phase.

fixup! chore(contracts): update `prepare` script in `package.json`