refactor(webapp): enable SnarkyJS by setting COOP and COEP headers
This change modifies the `next.config.js` file in the `webapp` package to add support for SnarkyJS on the web. The COOP (Cross-Origin-Opener-Policy) and COEP (Cross-Origin-Embedder-Policy) headers are set to `'same-origin'` and `'require-corp'` respectively. Additionally, a new `_headers` file is added to the `public` directory in the `webapp` package, which contains the same COOP and COEP headers.

This change enables SnarkyJS functionality and provides documentation on how to write a zkapp UI.

Link to documentation: [https://docs.minaprotocol.com/zkapps/how-to-write-a-zkapp-ui#enabling-coop-and-coep-headers](https://docs.minaprotocol.com/zkapps/how-to-write-a-zkapp-ui#enabling-coop-and-coep-headers)