chore: add analyze script to package.json and devDependencies
The `analyze` script has been added to the `scripts` section of `package.json`. This script is used to perform bundle analysis during the build process. Additionally, the `@next/bundle-analyzer` package has been added as a devDependency. This package is necessary for bundle analysis.

The changes have been made in the `packages/webapp/package.json` file and the `pnpm-lock.yaml` file.

Note: Ignore any changes to imports and requires.